Exploratory Factor Analysis

Wraps fa function from the psych package; see fa for more information.

addEFA(mctd, nfactors = length(unique(mctd$AnswerKey$Concept)),
  n.obs = nrow(mctd$Test.complete), rotate = "oblimin", fm = "minres",
  ...)

Arguments

mctd
Existing mcTestAnalysis data object

nfactors
Number of factors to extract, default is 1

n.obs
Number of observations used to find the correlation matrix if using a correlation matrix. Used for finding the goodness of fit statistics. Must be specified if using a correlaton matrix and finding confidence intervals.

rotate
"none", "varimax", "quartimax", "bentlerT", "equamax", "varimin", "geominT" and "bifactor" are orthogonal rotations. "Promax", "promax", "oblimin", "simplimax", "bentlerQ, "geominQ" and "biquartimin" and "cluster" are possible oblique transformations of the solution. The default is to do a oblimin transformation, although versions prior to 2009 defaulted to varimax. SPSS seems to do a Kaiser normalization before doing Promax, this is done here by the call to "promax" which does the normalization before calling Promax in GPArotation.

fm
Factoring method fm="minres" will do a minimum residual (OLS), fm="uls" differs very slightly from "minres" in that it minimizes the entire residual matrix using an OLS procedure. fm="wls" will do a weighted least squares (WLS) solution, fm="gls" does a generalized weighted least squares (GLS), fm="pa" will do the principal factor solution, fm="ml" will do a maximum likelihood factor analysis. fm="minchi" will minimize the sample size weighted chi square when treating pairwise correlations with different number of subjects per pair. fm ="minrank" will do a minimum rank factor analysis.

...
Passed to fa