Create MC Test Analysis Report

Creates a static PDF file containing a report on the properties of the test using Classic Test Theory and Item Response Theory.

createReport(answer_file = file.choose(), test_file = file.choose(),
  test_title = ifelse(interactive(), readline("Test title: "),
  "MC Test Analysis"), author = ifelse(interactive(),
  readline("Test author: "), Sys.info()["user"]), out_file = paste0("Report",
  " - ", test_title), out_path = getwd(), out_fmt = "pdf",
  open_file = TRUE, verbose = TRUE, report_options = list(), ...)

Arguments

answer_file
Path to the answer key data file

test_file
Path to the test results data file

test_title
Title of the test being analyzed

author
Name of the test administrator(s)

out_file
Desired name of report file when saved

out_path
Directory where report file should be saved, defaults to current working directory

out_fmt
Either 'pdf' or 'html', indicating format of report file. PDF files will be smaller and better typset, HTML files may be easier to navigate. Note that pdflatex is required to be installed. If pdflatex is not found, the report will fall back to HTML output.

open_file
Should the report file be opened automatically once generated?

verbose
If TRUE, notes about progress are written to the console.

report_options
A list of options that affect analysis parameters and results displayed inside the report. Report options include irt_model_choice, icc_group ("question" or "concept"), efa.nfactors, efa.cut, efa.rotate, efa.fm and distractor.pct.

...
Arguments passed to read.csv