MC Test Analysis Data

Load test and answer key data with loadTestData and loadAnswerKey. Creates an MC Test Analysis data object.

loadTestData(mctd = NULL, test_file, has_student_id = TRUE, ...)

loadAnswerKey(mctd = NULL, answer_file, ...)

loadAllData(answer_file = NULL, test_file = NULL, has_student_id = TRUE,
  force_load = FALSE, ...)

Arguments

mctd
Existing mcTestAnalysis data object

test_file
Path to the test results data file

has_student_id
Does the first column of the test data include a student identifier?

...
Arguments passed to read.csv

answer_file
Path to the answer key data file

force_load
Force calculation of test analysis items (may be slow). Default is FALSE as individual analysis will calculate needed analysis portions on the fly when called.

Functions

loadTestData: Reads test data from CSV or TSV file and applies basic preprocessing.

loadAnswerKey: Reads answer_key data from CSV or TSV file and applies basic preprocessing.

loadAllData: Read answer_key and test data from CSV or TSV files

Answer Key Data Format

The Answer Key file should contain four columns in the following order:

Question number, name or identifier: Eg. "Q1", "1", etc.

The correct answer for the question: Eg. "1", "A", etc.

A descriptive title for the question: Eg. "Taylor Function", "Tensor Flow", etc.

An identifier for the concept group to which the question belongs: Eg. "Taylor Series", "A", "Concept 1", etc.

Test Data Format

The test results data should contain as rows each student's response, with each question assigned a column. If the test results data contains student identifiers, these identifiers should be included in the first column, prior to the test answers. The MC Test Analysis Tool assumes that the question columns are in the same order as reported in the answers data.