Plot Discrimination Index vs Difficulty Index

Plots discrimination index, as calculated using the "conventional" formula, or via "PBCC", against difficulty index. Several parameters control settings related to the appearance of the final plot.

plotDiscriminationDifficulty(mctd, type = "conventional",
  show_labels = TRUE, hide_legend = TRUE, show_guidelines = TRUE,
  show_gridlines = TRUE, max_limits = "max_x")

Arguments

mctd
Existing mcTestAnalysis data object

type
One of "conventional", "pbcc", or "pbcc_modified"

show_labels
Should the question number be shown next to points? If ggrepel is installed it will be used to ensure that all point labels are legible on the plot.

hide_legend
Should the plot legend be hidden?

show_guidelines
Should recomended discrimination or difficulty indice ranges be shown on the plot as a dotted line?

show_gridlines
Should the plot include gridlines?

max_limits
Set x- and y-axis limits. One of "max_x" (x positive, y free), "max_y" (x free, y positive), "max_y+" (x free, y in [-1,1]), "max_all+" (x positive, y positive), "max_all" (x positive, y in [-1,1]), and NULL (all free).