Recommend review actions for each item

Recommends review actions for each item based on a set of heuristics or rules based on published guidelines.

recommendItemActions(mctd, include_columns = c("Title", "Concept",
  "Alpha WOI", "Difficulty", "Discrimination", "PBCC"),
  digits.round = getOption("digits"))

Arguments

mctd
Existing mcTestAnalysis data object

include_columns
Columns to include in recommendation table, can include any of the following: 'Title', 'Concept', 'Alpha WOI', 'Difficulty', 'Discrimination', 'PBCC'

digits.round
Round output to specified number of digits, defaults to digits option (see getOption("digits"))

Alpha

If Cronbach's Alpha for the test with the item deleted is less than the alpha coefficient for the whole test then the recommendation is to Keep the item.

Jorion

Following the recommendations in Jorion (2015), if the Difficulty Index is between 0.3 and 0.9, and the Discrimination Index is greater than 0.2, then the recommendation is to Keep the item.

Versatile

This recommendation is based on the Difficulty Index and PBCC and provides a range of recommendations from Remove to Review through Keep, favoring positive PBCC values near to or greater than 0.3 and higher difficulty values.

Stringent

If the Difficulty Index is between 0.3 and 0.9, and the point-biserial correlation coefficient is greater than 0.3, then the recommendation is to Keep the item.

References

Jorion, N., Gane, B. D., James, K., Schroeder, L., Dibello, L. V., & Pellegrino, J. W. (2015). An Analytic Framework for Evaluating the Validity of Concept Inventory Claims. Journal of Engineering Education, 104(4), 454-496. http://doi.org/10.1002/jee.20104

Sleeper, Rebecca. 2011. "Keep, Toss or Revise? Tips for Post-Exam Item Analysis." http://www.ttuhsc.edu/sop/administration/enhancement/documents/Sleeper_Handout.ppt (No longer available online)